Simplify initialization/reset sequences for compiled-code interface; revert previous change that caused disk-restore to fail on liarc.
Change raw-address operations (e.g. read_uuo_target) to work directly
with addresses rather than translating to and from objects.  This
translation is different for fasload.

The GC now has customizable operations for translating raw addresses
to and from objects.  This allows LOAD-BAND to override the read
translation so as to get the right behavior.

LOAD-BAND must relocate addresses if memory_base is different in the
band.  It also now reads pointers using the memory_base from the file,
and writes them using memory_base from the microcode.